+++ crema/__init__.py
@@ -5,4 +5,5 @@
 from .version import version as __version__
 
 from .exceptions import *
+from . import layers
 from . import models