+++ training/chords/01-prepare.py
@@ -50,9 +50,10 @@ def process_arguments(args):
 
 
 def make_pump(sr, hop_length, n_octaves):
-    p_feature = pumpp.feature.CQTMag(name='cqt',
-                                     sr=sr, hop_length=hop_length,
-                                     log=True, conv='tf', n_octaves=n_octaves)
+    p_feature = pumpp.feature.HCQTMag(name='cqt',
+                                      sr=sr, hop_length=hop_length,
+                                      harmonics=[1, 2, 3],
+                                      log=True, conv='tf', n_octaves=n_octaves)
 
     p_chord_tag = pumpp.task.ChordTagTransformer(name='chord_tag',
                                                  sr=sr, hop_length=hop_length,