+++ training/chords/02-train.py
@@ -16,6 +16,7 @@ from sklearn.model_selection import ShuffleSplit
 import pescador
 import librosa
 import crema.utils
+import crema.layers
 from jams.util import smkdirs
 
 OUTPUT_PATH = 'resources'
@@ -78,7 +79,7 @@ def make_sampler(max_samples, duration, pump, seed):
 
     n_frames = librosa.time_to_frames(duration,
                                       sr=pump['cqt'].sr,
-                                      hop_length=pump['cqt'].hop_length)[0]
+                                      hop_length=pump['cqt'].hop_length)
 
     return pump.sampler(max_samples, n_frames, random_state=seed)
 
@@ -89,7 +90,7 @@ def data_sampler(fname, sampler):
         yield datum
 
 
-def data_generator(working, tracks, sampler, k, augment=True, batch_size=32,
+def data_generator(working, tracks, sampler, k, augment=True, rate=8,
                    **kwargs):
     '''Generate a data stream from a collection of tracks and a sampler'''
 
@@ -106,34 +107,7 @@ def data_generator(working, tracks, sampler, k, augment=True, batch_size=32,
                 seeds.append(pescador.Streamer(data_sampler, fname, sampler))
 
     # Send it all to a mux
-    mux = pescador.Mux(seeds, k, **kwargs)
-
-    if batch_size == 1:
-        return mux
-    else:
-        return pescador.BufferedStreamer(mux, batch_size)
-
-
-def keras_tuples(gen, inputs=None, outputs=None):
-
-    if isinstance(inputs, six.string_types):
-        if isinstance(outputs, six.string_types):
-            # One input, one output
-            for datum in gen:
-                yield (datum[inputs], datum[outputs])
-        else:
-            # One input, multi outputs
-            for datum in gen:
-                yield (datum[inputs], [datum[o] for o in outputs])
-    else:
-        if isinstance(outputs, six.string_types):
-            for datum in gen:
-                yield ([datum[i] for i in inputs], datum[outputs])
-        else:
-            # One input, multi outputs
-            for datum in gen:
-                yield ([datum[i] for i in inputs],
-                       [datum[o] for o in outputs])
+    return pescador.StochasticMux(seeds, k, rate, **kwargs)
 
 
 def construct_model(pump):
@@ -153,16 +127,12 @@ def construct_model(pump):
                                    data_format='channels_last')(x_bn)
 
     # Second convolutional filter: a bank of full-height filters
-    conv2 = K.layers.Convolution2D(36, (1, int(conv1.shape[2])),
+    conv2 = K.layers.Convolution2D(12*6, (1, int(conv1.shape[2])),
                                    padding='valid', activation='relu',
                                    data_format='channels_last')(conv1)
 
     # Squeeze out the frequency dimension
-    def _squeeze(x):
-        import keras
-        return keras.backend.squeeze(x, axis=2)
-
-    squeeze = K.layers.Lambda(_squeeze)(conv2)
+    squeeze = crema.layers.SqueezeLayer(axis=2)(conv2)
 
     # BRNN layer
     rnn1 = K.layers.Bidirectional(K.layers.GRU(128,
@@ -267,21 +237,26 @@ def train(working, max_samples, duration, rate,
     gen_train = data_generator(working,
                                idx_train['id'].values, sampler, epoch_size,
                                augment=True,
-                               lam=rate,
-                               batch_size=batch_size,
-                               revive=True,
+                               rate=rate,
+                               mode='with_replacement',
                                random_state=seed)
 
-    gen_train = keras_tuples(gen_train(), inputs=inputs, outputs=outputs)
+    gen_train = pescador.maps.keras_tuples(pescador.maps.buffer_stream(gen_train(),
+                                                                       batch_size,
+                                                                       axis=0),
+                                           inputs=inputs,
+                                           outputs=outputs)
 
     gen_val = data_generator(working,
                              idx_val['id'].values, sampler, len(idx_val),
-                             augment=False,
-                             batch_size=batch_size,
-                             revive=True,
+                             augment=True,
+                             rate=rate,
+                             mode='with_replacement',
                              random_state=seed)
 
-    gen_val = keras_tuples(gen_val(), inputs=inputs, outputs=outputs)
+    gen_val = pescador.maps.keras_tuples(gen_val(),
+                                         inputs=inputs,
+                                         outputs=outputs)
 
     loss = {'chord_tag': 'sparse_categorical_crossentropy'}
     metrics = {'chord_tag': 'sparse_categorical_accuracy'}