+++ docs/changes.rst
@@ -3,6 +3,14 @@
 Release notes
 =============
 
+v0.2.0
+------
+Updated to work with newer upstream dependencies.
+
+Models should now deserialize properly across python versions.
+
+Chord model now uses viterbi decoding.
+
 v0.1.0
 ------
 Initial public release.